+++ demo/required_commands.go
@@ -12,5 +12,8 @@ func RequiredCommands() map[string]map[string]string {
 	cmds["cmctl"] = make(map[string]string)
 	cmds["cmctl"]["darwin"] = "brew install cmctl"
 
+	cmds["minio-mc"] = make(map[string]string)
+	cmds["cmctl"]["darwin"] = "brew install minio-mc"
+
 	return cmds
 }